Extract all exception definitions into a separate module to break an import cycle
